# docs/swagger.yaml
basePath: /
definitions:
  hotel-booking-app_internal_handler_dto.AddHotelDTO:
    properties:
      description:
        type: string
      location:
        type: string
      name:
        type: string
    required:
    - location
    - name
    type: object
  hotel-booking-app_internal_handler_dto.AddRoomDTO:
    properties:
      max_occupancy:
        type: integer
      price_per_night:
        type: number
      room_type:
        type: string
    type: object
  hotel-booking-app_internal_handler_dto.CreateReservationDTO:
    properties:
      check_in_date:
        type: string
      check_out_date:
        type: string
    required:
    - check_in_date
    - check_out_date
    type: object
  hotel-booking-app_internal_handler_dto.ErrorDTO:
    properties:
      message:
        type: string
    type: object
  hotel-booking-app_internal_handler_dto.HotelDTO:
    properties:
      description:
        type: string
      id:
        type: string
      location:
        type: string
      name:
        type: string
    type: object
  hotel-booking-app_internal_handler_dto.ReservationDTO:
    properties:
      check_in_date:
        type: string
      check_out_date:
        type: string
      id:
        type: string
      payment_status:
        type: string
      room:
        $ref: '#/definitions/hotel-booking-app_internal_handler_dto.RoomDTO'
      status:
        type: string
      user_id:
        type: string
    type: object
  hotel-booking-app_internal_handler_dto.ReturnIdDTO:
    properties:
      id:
        type: string
    type: object
  hotel-booking-app_internal_handler_dto.RoomDTO:
    properties:
      hotel:
        $ref: '#/definitions/hotel-booking-app_internal_handler_dto.HotelDTO'
      id:
        type: string
      max_occupancy:
        type: integer
      price_per_night:
        type: number
      room_type:
        type: string
    type: object
  hotel-booking-app_internal_handler_dto.SignInRequestDTO:
    properties:
      email:
        type: string
      password:
        maxLength: 255
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  hotel-booking-app_internal_handler_dto.SignInResponseDTO:
    properties:
      token:
        type: string
    type: object
  hotel-booking-app_internal_handler_dto.SignUpRequestDTO:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        maxLength: 255
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  hotel-booking-app_internal_handler_dto.SuccessDTO:
    properties:
      message:
        type: string
    type: object
  hotel-booking-app_internal_handler_dto.UpdateHotelDTO:
    properties:
      description:
        type: string
      location:
        type: string
      name:
        type: string
    type: object
  hotel-booking-app_internal_handler_dto.UpdateReservationStatusDTO:
    properties:
      status:
        type: string
    type: object
  hotel-booking-app_internal_handler_dto.UpdateRoomDTO:
    properties:
      max_occupancy:
        type: integer
      price_per_night:
        type: number
      room_type:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  title: Backend API
  version: "1.0"
paths:
  /hotel:
    post:
      consumes:
      - application/json
      parameters:
      - description: Hotel data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/hotel-booking-app_internal_handler_dto.AddHotelDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ReturnIdDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Add a new hotel
      tags:
      - Hotel
  /hotel/{id}:
    delete:
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.SuccessDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Delete a hotel by ID
      tags:
      - Hotel
    get:
      consumes:
      - application/json
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.HotelDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Get a hotel by ID
      tags:
      - Hotel
    put:
      consumes:
      - application/json
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      - description: Hotel data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/hotel-booking-app_internal_handler_dto.UpdateHotelDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.SuccessDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Update hotel by id
      tags:
      - Hotel
  /hotel/{id}/room:
    post:
      consumes:
      - application/json
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      - description: New room data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/hotel-booking-app_internal_handler_dto.AddRoomDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ReturnIdDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Add a new room to a hotel
      tags:
      - Room
  /hotel/{id}/room/all:
    get:
      consumes:
      - application/json
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hotel-booking-app_internal_handler_dto.RoomDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Get all rooms for a hotel
      tags:
      - Room
  /hotel/{id}/room/free:
    get:
      consumes:
      - application/json
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hotel-booking-app_internal_handler_dto.RoomDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Get all free rooms for a hotel
      tags:
      - Room
  /hotel/all:
    get:
      consumes:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hotel-booking-app_internal_handler_dto.HotelDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Get all hotels
      tags:
      - Hotel
  /reservation/:id/cancel:
    get:
      consumes:
      - application/json
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.SuccessDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Cancel a user's reservation
      tags:
      - Reservation
  /reservation/:id/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      - description: Reservation status data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/hotel-booking-app_internal_handler_dto.UpdateReservationStatusDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.SuccessDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Update the status of a reservation
      tags:
      - Reservation
  /reservation/all:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ReservationDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Get all reservations for the authenticated user
      tags:
      - Reservation
  /reservation/all/manager:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ReservationDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Get all reservations (this request for the manager)
      tags:
      - Reservation
  /room/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.SuccessDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Delete a room from a hotel
      tags:
      - Room
    get:
      consumes:
      - application/json
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.RoomDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Get a room by ID
      tags:
      - Room
    put:
      consumes:
      - application/json
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      - description: Room data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/hotel-booking-app_internal_handler_dto.UpdateRoomDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.SuccessDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Update a room in a hotel
      tags:
      - Room
  /room/{id}/reserve:
    post:
      consumes:
      - application/json
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      - description: Reservation data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/hotel-booking-app_internal_handler_dto.CreateReservationDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ReturnIdDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Create a reservation for a room in a hotel
      tags:
      - Reservation
  /sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: User credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/hotel-booking-app_internal_handler_dto.SignInRequestDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.SignInResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Sign in to account
      tags:
      - Authentication
  /sign-up:
    post:
      consumes:
      - application/json
      parameters:
      - description: User credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/hotel-booking-app_internal_handler_dto.SignUpRequestDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ReturnIdDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hotel-booking-app_internal_handler_dto.ErrorDTO'
      summary: Sign up into account
      tags:
      - Authentication
swagger: "2.0"